server/requirements.txt
# Core Dependencies
python-dotenv>=1.0.1
numpy
# --- Web Framework & Server ---
fastapi>=0.110.0
uvicorn>=0.15.0
starlette>=0.27.0

# --- Database ---
sqlalchemy[asyncio]==1.4.46
asyncpg==0.27.0
aiosqlite==0.20.0

# --- Authentication & Security ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# --- Data Processing & Validation ---
pydantic>=2.11.2,<3.0.0
pydantic-settings>=2.4.0

# --- LangChain Core ---
# langchain>=0.3.0
# langchain-core>=0.3.0 
# langchain-community>=0.3.0
# langchain-openai>=0.2.0
# langchain-text-splitters>=0.3.0

# # --- AI/ML Models ---
# openai>=1.40.0,<2.0.0
# google-generativeai==0.3.1

# # --- Vector Stores ---
# weaviate-client==3.25.2

# # --- Document Loaders ---
# unstructured>=0.10.30
# unstructured-client>=0.42.4

# --- AI Tools ---
# transformers>=4.30.0  # Version with pre-built wheels for Windows
# For Windows, we'll use a more compatible version of faster-whisper
# faster-whisper>=0.9.0; platform_system != 'Windows'
# faster-whisper>=1.0.0; platform_system == 'Windows' and python_version < '3.13'
# --- Web & API ---
python-socketio==5.8.0
httpx==0.27.0

# --- Development (Optional) ---
# ipywidgets==8.1.2
# jupyter==1.0.0
# pygments==2.17.2
# python-magic==0.4.27
# pygraphviz==1.13  # Requires system-level graphviz installation

# langchain-google-genai
